refactor(types): split ids.ts by ID-domain
Bean: types-vt49
